# Django 4.x and other package upgrades *Beta testing needed*