--- a/backend/Dockerfile
+++ b/backend/Dockerfile
@@ -7,5 +7,3 @@ RUN pip install -U pip
 ADD ./requirements.txt /app
 RUN pip install -r requirements.txt
 ADD ./ /app
-
-ENTRYPOINT ["/app/entrypoint.sh"]