--- a/backend/Dockerfile.base
+++ b/backend/Dockerfile.base
@@ -1,4 +1,4 @@
-FROM python:3.9-alpine3.14
+FROM python:3.11-alpine3.18
 
 ENV DEBIAN_FRONTEND=noninteractive
 
@@ -6,7 +6,7 @@ WORKDIR /app
 EXPOSE 3334
 RUN apk -U add bash vim ffmpeg postgresql-libs git openjpeg tiff
 RUN apk add --virtual .build-deps g++ musl-dev postgresql-dev zlib-dev jpeg-dev libffi-dev openjpeg-dev tiff-dev \
-    patch curl-dev python3-dev libressl-dev build-base freetype-dev lcms2-dev tk-dev tcl-dev  # Needed by /moonraker/scripts/moonraker-requirements.txt
+    patch curl-dev python3-dev build-base freetype-dev lcms2-dev tk-dev tcl-dev  # Needed by /moonraker/scripts/moonraker-requirements.txt
 
 # Checkout moonraker and lock the version so that we can use the metadata parser
 RUN git clone https://github.com/Arksine/moonraker.git /moonraker
@@ -15,11 +15,7 @@ ENV PYTHONPATH="${PYTHONPATH}:/moonraker/moonraker"
 
 RUN pip install --upgrade pip
 
-ADD requirements.txt .
-RUN CRYPTOGRAPHY_DONT_BUILD_RUST=1 pip install -r requirements.txt
-
-RUN pip install tornado==6.2.0 inotify-simple==1.3.5
-# Apparent newer version of importlib-metadata will cause django to fail
-RUN pip install importlib-metadata==4.13.0
+ADD requirements-4.x.txt .
+RUN CRYPTOGRAPHY_DONT_BUILD_RUST=1 pip install -r requirements-4.x.txt
 
 RUN apk --purge del .build-deps