--- a/backend/Dockerfile.base
+++ b/backend/Dockerfile.base
@@ -1,4 +1,4 @@
-FROM python:3.11-alpine3.18
+FROM python:3.10-alpine3.18
 
 ENV DEBIAN_FRONTEND=noninteractive
 
@@ -6,7 +6,7 @@ WORKDIR /app
 EXPOSE 3334
 RUN apk -U add bash vim ffmpeg postgresql-libs git openjpeg tiff
 RUN apk add --virtual .build-deps g++ musl-dev postgresql-dev zlib-dev jpeg-dev libffi-dev openjpeg-dev tiff-dev \
-    cmake patch curl-dev python3-dev build-base freetype-dev lcms2-dev tk-dev tcl-dev  # Needed by /moonraker/scripts/moonraker-requirements.txt
+    openblas-dev cmake patch curl-dev python3-dev build-base freetype-dev lcms2-dev tk-dev tcl-dev  # Needed by /moonraker/scripts/moonraker-requirements.txt
 
 # Checkout moonraker and lock the version so that we can use the metadata parser
 RUN git clone https://github.com/Arksine/moonraker.git /moonraker