--- a/backend/Pipfile
+++ b/backend/Pipfile
@@ -5,7 +5,7 @@ name = "pypi"
 
 [packages]
 ipdb = "*"
-django = "==4.*"
+django = "==4.1.*"
 gunicorn = "*"
 dj-database-url = "*"
 psycopg2-binary = "*"