--- a/backend/Pipfile
+++ b/backend/Pipfile
@@ -5,7 +5,7 @@ name = "pypi"
 
 [packages]
 ipdb = "*"
-django = "==4.1.*"
+django = "==4.0.*"
 gunicorn = "*"
 dj-database-url = "*"
 psycopg2-binary = "*"
@@ -36,7 +36,7 @@ django-jstemplate = "*"
 pytelegrambotapi = "*"
 # Fork of django_channels_presence adds django 4.x compatibility
 # See https://github.com/mitmedialab/django-channels-presence/issues/23
-django-channels-presence = {ref = "24c15f229372a16c7e8d6e160de89f4817d96c3f", git = "https://github.com/ApexArray/django-channels-presence"}
+django-channels-presence = {ref = "24c15f229372a16c7e8d6e160de89f4817d96c3f", git = "https://github.com/TheSpaghettiDetective/django-channels-presence"}
 backoff = "*"
 django-webpack-loader = "*"
 django-qr-code = "*"
@@ -51,7 +51,7 @@ sentry-sdk = "*"
 nplusone = "*"
 colorlog = "*"
 packaging = "*"
-numpy = "*"
+numpy = "1.2.0"
 django-debug-toolbar = "*"
 inotify-simple = "*"
 Twisted = {extras = ["tls", "http2"], version = "*" }
@@ -59,4 +59,4 @@ Twisted = {extras = ["tls", "http2"], version = "*" }
 [dev-packages]
 
 [requires]
-python_version = "3.11"
+python_version = "3.10"