--- a/backend/api/viewsets.py
+++ b/backend/api/viewsets.py
@@ -213,7 +213,7 @@ class PrintViewSet(
 
     def list(self, request):
         queryset = self.get_queryset().prefetch_related('printshotfeedback_set'
-            ).select_related('printer', 'g_code_file',
+            ).select_related('printer', 'g_code_file', 'printer__user'
             )
 
         sorting = request.GET.get('sorting', 'date_desc')
@@ -575,7 +575,7 @@ class PrintShotFeedbackViewSet(mixins.RetrieveModelMixin,
 
         qs = PrintShotFeedback.objects.filter(
             print__user=self.request.user
-        )
+        ).select_related('user')
 
         if print_id:
             qs = qs.filter(print_id=print_id)