--- a/backend/api/ws_routing.py
+++ b/backend/api/ws_routing.py
@@ -1,5 +1,4 @@
-from django.urls import path
-from django.conf.urls import re_path
+from django.urls import path, re_path
 
 from . import consumers