--- a/backend/app/__init__.py
+++ b/backend/app/__init__.py
@@ -1,4 +1,3 @@
 from config import celery_app
 
-default_app_config = 'app.apps.WebAppConfig'
 __all__ = ('celery_app',)