--- a/backend/app/apps.py
+++ b/backend/app/apps.py
@@ -4,7 +4,3 @@ from django.apps import AppConfig
 
 class WebAppConfig(AppConfig):
     name = 'app'
-
-
-class CustomRoomsConfig(RoomsConfig):
-    name = 'channels_presence'