--- a/backend/app/middleware.py
+++ b/backend/app/middleware.py
@@ -70,10 +70,12 @@ def fix_tunnelv2_apple_cache(get_response):
 class RefreshSessionMiddleware(SessionMiddleware):
     def process_response(self, request, response):
         session = request.session
+        if settings.SESSION_COOKIE_REFRESH_INTERVAL >= settings.SESSION_COOKIE_AGE:
+            return ValueError("SESSION_COOKIE_REFRESH_INTERVAL must be less than SESSION_COOKIE_AGE")
         if not (session.is_empty() or session.get_expire_at_browser_close()):
-            expires_at_ts = session.get('_session_expire_at_ts', None)
             now_ts = int(time.time())
-            refresh_at_ts = expires_at_ts - (settings.SESSION_COOKIE_AGE - settings.SESSION_COOKIE_REFRESH_INTERVAL)
+            expires_at_ts = session.get('_session_expire_at_ts', None)
+            refresh_at_ts = expires_at_ts or now_ts - (settings.SESSION_COOKIE_AGE - settings.SESSION_COOKIE_REFRESH_INTERVAL)
             if expires_at_ts is None or now_ts >= refresh_at_ts:
                 # This will set modified flag and update the cookie expiration time
                 session['_session_expire_at_ts'] = now_ts + settings.SESSION_COOKIE_AGE