--- a/backend/app/middleware.py
+++ b/backend/app/middleware.py
@@ -16,6 +16,7 @@ from lib.tunnelv2 import OctoprintTunnelV2Helper
 
 import logging
 
+LOGGER = logging.getLogger()
 
 class TSDWhiteNoiseMiddleware(WhiteNoiseMiddleware):
 
@@ -65,66 +66,19 @@ def fix_tunnelv2_apple_cache(get_response):
 
     return middleware
 
-
-# https://stackoverflow.com/questions/2116860/django-session-cookie-domain-with-multiple-domains
-# https://ittone.ma/ittone/django-session_cookie_domain-with-multiple-domains/
-class SessionHostDomainMiddleware(SessionMiddleware):
-
+# https://stackoverflow.com/questions/64466605/django-how-to-get-the-time-until-a-cookie-expires
+class RefreshSessionMiddleware(SessionMiddleware):
     def process_response(self, request, response):
-        """
-        If request.session was modified, or if the configuration is to save the
-        session every time, save the changes and set a session cookie or delete
-        the session cookie if the session has been emptied.
-        """
-        try:
-            accessed = request.session.accessed
-            modified = request.session.modified
-            empty = request.session.is_empty()
-        except AttributeError:
-            pass
-        else:
-            host = request.get_host().split(':')[0]
-            # First check if we need to delete this cookie.
-            # The session should be deleted only if the session is entirely empty
-            if settings.SESSION_COOKIE_NAME in request.COOKIES and empty:
-                response.delete_cookie(
-                    settings.SESSION_COOKIE_NAME,
-                    path=settings.SESSION_COOKIE_PATH,
-                    domain=host,
-                    samesite=settings.SESSION_COOKIE_SAMESITE,
-                )
-            else:
-                if accessed:
-                    patch_vary_headers(response, ('Cookie',))
-                if (modified or settings.SESSION_SAVE_EVERY_REQUEST) and not empty:
-                    if request.session.get_expire_at_browser_close():
-                        max_age = None
-                        expires = None
-                    else:
-                        max_age = request.session.get_expiry_age()
-                        expires_time = time.time() + max_age
-                        expires = http_date(expires_time)
-                    # Save the session data and refresh the client cookie.
-                    # Skip session save for 500 responses, refs #3881.
-                    if response.status_code != 500:
-                        try:
-                            request.session.save()
-                        except UpdateError:
-                            raise SuspiciousOperation(
-                                "The request's session was deleted before the "
-                                "request completed. The user may have logged "
-                                "out in a concurrent request, for example."
-                            )
-                        response.set_cookie(
-                            settings.SESSION_COOKIE_NAME,
-                            request.session.session_key, max_age=max_age,
-                            expires=expires, domain=host,
-                            path=settings.SESSION_COOKIE_PATH,
-                            secure=settings.SESSION_COOKIE_SECURE or None,
-                            httponly=settings.SESSION_COOKIE_HTTPONLY or None,
-                            samesite=settings.SESSION_COOKIE_SAMESITE,
-                        )
-        return response
+        session = request.session
+        if not (session.is_empty() or session.get_expire_at_browser_close()):
+            expires_at_ts = session.get('_session_expire_at_ts', None)
+            now_ts = int(time.time())
+            refresh_at_ts = expires_at_ts - (settings.SESSION_COOKIE_AGE - settings.SESSION_COOKIE_REFRESH_INTERVAL)
+            if expires_at_ts is None or now_ts >= refresh_at_ts:
+                # This will set modified flag and update the cookie expiration time
+                session['_session_expire_at_ts'] = now_ts + settings.SESSION_COOKIE_AGE
+        return super().process_response(request, response)
+
 
 def check_admin_ip_whitelist(get_response):
     try: