--- a/backend/app/templatetags/vue.py
+++ b/backend/app/templatetags/vue.py
@@ -13,11 +13,11 @@ TEMPLATES = {
 }
 
 
-@register.simple_tag
-def bundle(bundle_name, extension=None, config='DEFAULT', attrs=''):
+@register.simple_tag(takes_context=True)
+def bundle(context, bundle_name, extension=None, config='DEFAULT', attrs=''):
     if settings.WEBPACK_LOADER_ENABLED:
         return render_bundle(
-            bundle_name, extension=extension, config=config, attrs=attrs)
+            context, bundle_name, extension=extension, config=config, attrs=attrs)
     else:
         tags = []
         extensions = (extension, ) if extension is not None else ('js', 'css')
@@ -36,12 +36,12 @@ def bundle(bundle_name, extension=None, config='DEFAULT', attrs=''):
         return mark_safe("\n".join(tags))
 
 
-@register.simple_tag
-def bundlechunk(bundle_name, extension=None, config='DEFAULT', attrs=''):
+@register.simple_tag(takes_context=True)
+def bundlechunk(context, bundle_name, extension=None, config='DEFAULT', attrs=''):
     """ Chunks are generated only for production and must be loaded only
         when loader is disabled
     """
     if not settings.WEBPACK_LOADER_ENABLED:
-        return bundle(bundle_name, extension=extension, config=config,
+        return bundle(context, bundle_name, extension=extension, config=config,
                       attrs=attrs)
     return ""