--- a/backend/config/settings.py
+++ b/backend/config/settings.py
@@ -59,7 +59,6 @@ INSTALLED_APPS = [
     'simple_history',
     'widget_tweaks',
     'rest_framework',
-    'bootstrap_pagination',
     'jstemplate',
     'pushbullet',
     'corsheaders',