--- a/backend/config/settings.py
+++ b/backend/config/settings.py
@@ -113,7 +113,7 @@ if DEBUG:
     INTERNAL_IPS = [ip[: ip.rfind(".")] + ".1" for ip in ips] + ["127.0.0.1", "10.0.2.2"]
 
     # Add nplusone
-    INSTALLED_APPS += 'nplusone.ext.django'
+    INSTALLED_APPS.append('nplusone.ext.django')
     MIDDLEWARE.insert(gzip_index+1, 'nplusone.ext.django.NPlusOneMiddleware')
 
 ROOT_URLCONF = 'config.urls'