--- a/backend/entrypoint.sh
+++ b/backend/entrypoint.sh
@@ -2,5 +2,6 @@
 
 # Ensure all migrations are applied
 python manage.py migrate
+python manage.py collectstatic -v 2 --noinput
 
 exec "$@"