--- a/backend/entrypoint.sh
+++ b/backend/entrypoint.sh
@@ -2,6 +2,5 @@
 
 # Ensure all migrations are applied
 python manage.py migrate
-python manage.py collectstatic -v 2 --noinput
 
 exec "$@"