--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -73,4 +73,4 @@ services:
 
   redis:
     restart: unless-stopped
-    image: redis:5.0-alpine
+    image: redis:7.2-alpine