… .github/workflows/build-images.yaml
… backend/Pipfile.lock
… backend/app/management/commands/init_db_with_test_data.py
… backend/app/storage.py
… backend/app/templates/layout.html
… backend/requirements-3.x.txt
… backend/requirements-4.x.txt
… backend/requirements-base.txt
… backend/requirements.txt
… docker-compose-build-base.yml
… docker-compose-dev.yml
… frontend/builds/frontend/js/chunk-2d0d5f72.js
… frontend/builds/frontend/js/chunk-2d0d5f72.js.map
… frontend/package.json
… frontend/yarn.lock
… load_testing/docker-compose-load-testing.yml
… load_testing/locustfile_ml_api.py
… yarn.lock