# Overview

This PR bumps Python and its packages to their most current release, bringing numerous improvements to performance and security. Most notably:
- Python 3.7 -> 3.11
- Django 2.2 -> 4.2
- Django Channels / Daphne 2.5 -> 4.0
- Celery 4 -> 5
- Redis 5.0 -> 7.2

Since I was already refactoring much of the codebase for compatibility with the new versions, I also took the opportunity to make some other improvements:

# Performance optimizations
- Disable SESSION_SAVE_EVERY_REQUEST, and refresh the session every 24 hours instead with custom `RefreshSessionMiddleware` middleware.
- Use [cached sessions](https://docs.djangoproject.com/en/4.2/topics/http/sessions/#using-cached-sessions) with Django's `cached_db` SESSION_ENGINE. 
- Only enable debugging packages if DEBUG=True (nplusone, django debug toolbar)
- Fixed an n+1 query in the PrinterViewSet

# Docker optimizations
- Optimize Dockerfile to take advantage of [layer caching](https://docs.docker.com/build/cache/)
- Use `entrypoint.sh` instead of chained CMD
- Add `docker-compose-build-base.yml` to make it easier to build the base image
- Add `docker-compose-dev.yml` to make it easier to start dev environment
  - Runs dev server using `manage.py runserver` instead of `daphne` (hot reloads server after changing a file)
  - DEBUG=True by default
  - Skip entrypoint.sh for container start (does not run migrate or collecstatic)

# Other changes
- Use pipenv for dependency management
  - Run `cd backend && pipenv lock && pipenv requirements > requirements.txt` to generate a new requirements.txt file
- Optimized GitHub workflows in `build-images.yaml` (these changes made sense to me, but I won't be offended if you don't want to include them in this PR).
- Bump `webpack-bundle-tracker` package in frontend (for compatibility with new `django-webpack-loader` version)
- Optimized Django test suite (tests now complete in ~1s instead of ~35s)
  - Note that some tests are currently failing, but that was the case even before the package upgrades
- Register ALL models in Django admin if DEBUG=True
- Add optional locust configs under `./load_testing/`
- Add `init_db_with_test_data` management command. 
  - This command flushes the current database (deletes all data), and adds a test user and printer (primarily used for setting up a load testing environment)

# Upgrade instructions

**NOTE:** This upgrade should be reversible, but it's still good practice to [backup your database](https://www.obico.io/docs/server-guides/ops/#backup-database) before upgrading.

Simply checkout this branch and run `docker compose up -d`.

# PR Status

I will track any bugs or development tasks here.

I'm using this branch in my live environment and everything seems to be working smoothly, but more testing is needed. Please let me know if you're able to help with beta testing!