@ApexArray 2023-10-26
# Load testing results

Adding some quick load test results.

## Test conditions
- All relevant settings and environment details were identical, except for the package versions and refactored code.
- Locust ran on a separate computer to minimize its performance impact
- Initialized fresh database with `init_db_with_test_data` management command
- Tested following endpoints (weighted)
  - 50% `/api/v1/octo/pic/`
  - 16.7% `/api/v1/printers/`
  - 33.3% `/api/v1/printers/1/`
- Strictly measures performance within the web container
  - No active print, so pics were not submitted to ml_api 

## Server info:
- Debian 11 (bullseye) VM
- 2x vCPU E5-2667 v2 @ 3.30GHz
- 24GB RAM
- SSD Storage
- Docker 24.0.1
- PostgreSQL 14.2 (Unix socket)

## Results

### Old
~20 rps
50th percentile latency: 5300ms
95th percentile latency: 5600ms
![image](https://github.com/TheSpaghettiDetective/obico-server/assets/91706106/93af095f-df11-4485-94ef-bdf6972939c6)


### New
40-50 rps (100%+ increase)
50th percentile latency: 1700ms (~300% faster)
95th percentile latency: 3300ms (~170% faster)
![image](https://github.com/TheSpaghettiDetective/obico-server/assets/91706106/6160a541-d73c-498a-bfd0-797f55d1fe77)

## Full test reports
[locust_reports.zip](https://github.com/TheSpaghettiDetective/obico-server/files/13182448/locust_reports.zip)